requirements.txt
# Base requirements (install inside a virtual environment)
Flask>=2.0
Flask-Cors
python-dotenv==1.0.1
requests==2.31.0
notion-client==2.0.0

# Scheduler
Flask-APScheduler>=1.12.0

# AI and Transcription
google-generativeai>=0.3.0
openai-whisper

# Optional (recommended) for CPU-only installs of Whisper
# Install separately: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu

# Slack Integration
# Install slack_sdk inside a virtual environment
slack_sdk

# Web Server
# Install gunicorn and rapidfuzz inside a virtual environment
gunicorn
rapidfuzz